requirements.txt
PyQRCode==1.2.1
Pillow==8.4.0
qrcode==7.3.1
